lambda/requirements.txt
boto3>=1.28.57
botocore>=1.31.57
python-pptx==0.6.18